+++ lib/ui/page/call/settings/controller.dart
@@ -22,8 +22,8 @@ import 'package:medea_jason/medea_jason.dart';
 import '/domain/model/application_settings.dart';
 import '/domain/model/ongoing_call.dart';
 import '/domain/repository/settings.dart';
-import '/util/custom_scroll_controller.dart';
 import '/util/obs/obs.dart';
+import '/util/scroll_controller.dart';
 
 export 'view.dart';
