+++ lib/ui/page/home/page/chat/info/controller.dart
@@ -34,8 +34,8 @@ import '/l10n/l10n.dart';
 import '/provider/gql/exceptions.dart';
 import '/routes.dart';
 import '/ui/widget/text_field.dart';
-import '/util/custom_scroll_controller.dart';
 import '/util/message_popup.dart';
+import '/util/scroll_controller.dart';
 import '/util/web/web_utils.dart';
 
 export 'view.dart';
