+++ lib/ui/page/home/page/my_profile/controller.dart
@@ -42,8 +42,8 @@ import '/l10n/l10n.dart';
 import '/provider/gql/exceptions.dart';
 import '/routes.dart';
 import '/ui/widget/text_field.dart';
-import '/util/custom_scroll_controller.dart';
 import '/util/message_popup.dart';
+import '/util/scroll_controller.dart';
 
 export 'view.dart';
