+++ lib/ui/page/home/tab/menu/confirm/controller.dart
@@ -24,7 +24,7 @@ import '/domain/service/my_user.dart';
 import '/l10n/l10n.dart';
 import '/provider/gql/exceptions.dart' show UpdateUserPasswordException;
 import '/ui/widget/text_field.dart';
-import '/util/custom_scroll_controller.dart';
+import '/util/scroll_controller.dart';
 
 /// Possible [ConfirmLogoutView] flow stage.
 enum ConfirmLogoutViewStage {
