+++ lib/ui/page/home/tab/menu/status/controller.dart
@@ -26,7 +26,7 @@ import '/domain/model/user.dart';
 import '/domain/service/my_user.dart';
 import '/l10n/l10n.dart';
 import '/ui/widget/text_field.dart';
-import '/util/custom_scroll_controller.dart';
+import '/util/scroll_controller.dart';
 
 /// Controller of a [StatusView].
 class StatusController extends GetxController {
